Refactored unstable_renderSubtree into ReactDOM.createPortal
unstable_renderSubtree seems to require a mounted component, but moving it into componentDidMount still produces a failure to embed the field in dom. 

Perhaps someone with more React knowledge can explain why we can't just render the field by the normal means, ie. the render() component override? 
Am trying to stick as close to original author's code as possible. Thus... ReactDOM.createPortal.

If there is a better way to accomplish this, please submit a PR or write up an issue about it. Thanks!

Does it work? Yes.
Is it "right"? I don't know.

See Issue #1 , regarding React 16 support.
P.S.   the accounts-material-ui companion package had to be updated as well to support 16.